De-scope some ``session`` scoped fixtures
